requirements.txt
langchain
openai
google-api-python-client
python-dotenv
wget
PyPDF2
spacy
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.4.0/fr_core_news_sm-3.4.0-py3-none-any.whl